airflow/requirements.txt
yfinance
pandas
requests
google-cloud-storage
google-cloud-bigquery
dbt-bigquery